[BE] Refactor : dto 폴더 생성
