+++ BE/src/users/users.controller.ts
@@ -14,10 +14,10 @@ import {
   Headers,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { CreateUserDto } from './createUser.dto';
+import { CreateUserDto } from './dto/createUser.dto';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { MultiPartBody } from 'src/common/decorator/multiPartBody.decorator';
-import { UpdateUsersDto } from './usersUpdate.dto';
+import { UpdateUsersDto } from './dto/usersUpdate.dto';
 import { AuthGuard } from 'src/common/guard/auth.guard';
 import { UserHash } from '../common/decorator/auth.decorator';
 import { FileSizeValidator } from '../common/files.validator';
